Handle dotenv import failure gracefully
This commit introduces a try-except block around the dotenv import to prevent the application from crashing if the dotenv package is not installed. Now, if the import fails, the application will continue to run without loading the environment variables from a .env file. This change improves the robustness of the application startup process.